Use an upstream instead of a proxy (#10)
Much of this code is following what the ngx_http_proxy_module code does
to initialize an upstream and use it.

Once the upstream configuration is in place and an upstream object has
been created, our responsibilities are:
* Intercept subrequest that we want to send to this upstream instead of
letting the default handler take it. I re-used the same phase handler,
but it might be better to hook into a different part of the stack? This
works, though.
* Format a HTTP request and add it to the outgoing list of buffers in
the upstream.
* Parse the response as HTTP and set the appropriate fields in the
upstream's header object.

Co-authored-by: Jed Liu <liujed@users.noreply.github.com>