Run integration tests in CircleCI (#16)
This PR adds a test step that downloads and runs the Nginx integration
tests, and saves the module and test results as artifacts.

The plan is to run the full sets of builds and tests only on a newly
created "release" branch but this demos running the tests on any branch
to see if it actually works.

There's a test failure here which does not occur locally; we might need
to merge in some of the open PRs to get a pass. Or it might be a race
condition.

Also switches to building the module dynamically.